Move parser loop and conditional code to separate files
Most of this code has been moved out of parser.c into parser_cond.*
and parser_loops.*.
